alignment matrix+preprocessing
The allignment matrix is created, sometimes it will give a warning, but
I think it is save to ignore.
In combine alignment I also created new en_50000/pt_50000 files.
Sometimes sentences are merged with a previous or next sentence, which
creates a bad alignment and one sentence aligning to an empty line. So
these with their surrounding sentences are removed